fix(uninstall): fix issues with uninstalling manifest-only drivers (#5)
This primarily fixes a bug with dbc uninstall for manifest-only drivers
where the extra folder we install alongside the driver manifest wasn't
being removed. I also factored out the common logic related to removing
the driver shared library and added some defensive checks to limit the
impact of a maliciously-crafted driver manifest.

Ref #15

---------

+++ config/dirs_windows.go
@@ -5,7 +5,6 @@ package config
 import (
 	"errors"
 	"fmt"
-	"io/fs"
 	"log"
 	"maps"
 	"os"
@@ -275,40 +274,20 @@ func CreateManifest(cfg Config, driver DriverInfo) (err error) {
 }
 
 func UninstallDriver(cfg Config, info DriverInfo) error {
-	k, err := registry.OpenKey(cfg.Level.key(), regKeyADBC, registry.WRITE)
-	if err != nil {
-		return err
-	}
-	defer k.Close()
-
-	if err := registry.DeleteKey(k, info.ID); err != nil {
-		return fmt.Errorf("failed to delete driver registry key: %w", err)
-	}
-
-	if info.Source == "dbc" {
-		for sharedPath := range info.Driver.Shared.Paths() {
-			if err := os.RemoveAll(filepath.Dir(sharedPath)); err != nil {
-				// Ignore only when not found. This supports manifest-only drivers.
-				// TODO: Come up with a better mechanism to handle manifest-only drivers
-				// and remove this continue when we do
-				if errors.Is(err, fs.ErrNotExist) {
-					continue
-				}
-				return fmt.Errorf("error removing driver %s: %w", info.ID, err)
-			}
+	if cfg.Level != ConfigEnv {
+		k, err := registry.OpenKey(cfg.Level.key(), regKeyADBC, registry.WRITE)
+		if err != nil {
+			return err
 		}
-	} else {
-		for sharedPath := range info.Driver.Shared.Paths() {
-			if err := os.Remove(sharedPath); err != nil {
-				// Ignore only when not found. This supports manifest-only drivers.
-				// TODO: Come up with a better mechanism to handle manifest-only drivers
-				// and remove this continue when we do
-				if errors.Is(err, fs.ErrNotExist) {
-					continue
-				}
-				return fmt.Errorf("error removing driver %s: %w", info.ID, err)
-			}
+		defer k.Close()
+
+		if err := registry.DeleteKey(k, info.ID); err != nil {
+			return fmt.Errorf("failed to delete driver registry key: %w", err)
 		}
+	}
+
+	if err := UninstallDriverShared(cfg, info); err != nil {
+		return fmt.Errorf("failed to delete driver shared object: %w", err)
 	}
 
 	return nil
